Created a models subdirectory that now holds the GaussianProcess, and renamed static_inspection.h to traits.h.
